Adjustment on tokens and token utilities
