+++ test/src/factory/mix_provider_data_test.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mix/mix.dart';
+import 'package:mix/src/theme/token_resolver.dart';
 
 import '../../helpers/testing_utils.dart';
